cleanup examples app.js - abstract moving parts
