+++ examples/src/theme.js
@@ -1,0 +1,58 @@
+module.exports = {
+	// wrapper
+	wrapper: { background: 'rgba(255, 255, 255, 0.9)' },
+
+	// arrows
+	arrow: {
+		backgroundColor: 'rgba(255, 255, 255, 0.8)',
+		fill: '#222',
+		opacity: 0.6,
+		transition: 'opacity 200ms',
+
+		':hover': {
+			opacity: 1,
+		},
+	},
+	arrow__size__medium: {
+		borderRadius: 40,
+		height: 40,
+		marginTop: -20,
+
+		'@media (min-width: 768px)': {
+			height: 60,
+			padding: 15,
+		},
+	},
+	arrow__direction__left: { marginLeft: 10 },
+	arrow__direction__right: { marginRight: 10 },
+
+	// header
+	header: {
+		background: 'none',
+	},
+	close: {
+		fill: '#D40000',
+		opacity: 0.6,
+		transition: 'all 200ms',
+
+		':hover': {
+			opacity: 1,
+		},
+	},
+
+	// footer
+	footer: {
+		background: 'hsla(0, 0%, 100%, 0.66)',
+		color: 'black',
+	},
+	footerCount: {
+		color: 'rgba(0, 0, 0, 0.6)',
+	},
+
+	// thumbnails
+	thumbnail: {
+	},
+	thumbnail__active: {
+		boxShadow: '0 0 0 2px #00D8FF',
+	},
+};
